ajout login/logout

+++ src/Controller/DevisController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Devis;
 use App\Form\DevisType;
 use App\Repository\DevisRepository;
+use App\Service\UserStatusService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -12,9 +13,21 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class DevisController extends AbstractController
 {
+    private $statusService;
+
+    public function __construct(UserStatusService $statusService) {
+        $this->statusService = $statusService;
+    }
+
     #[Route('/devis', name: 'app_devis')]
     public function Devis(DevisRepository $DevisRepository)
     {
+        $status = $this->statusService->getStatus();
+
+        if ($status == true) {
+            return $this->render('unauthorized.html.twig');
+        }
+        
         $Devis = $DevisRepository->findAll();
         return $this->render('devis/listesDevis.html.twig', [
             'devis' => $Devis,
@@ -25,7 +38,7 @@ class DevisController extends AbstractController
     #[Route('/ajouterdevis', name: 'app_ajouter_devis')]
     public function ajouterDevis(Request $request, EntityManagerInterface $em)
     {   
-        $status = true;
+        $status = $this->statusService->getStatus();
 
         if ($status == true) {
             return $this->render('unauthorized.html.twig');
@@ -53,6 +66,12 @@ class DevisController extends AbstractController
     #[Route("/modifierdevis/{id<\d+>}", name: "app_modifier_devis")]
     public function modifierDevis(Request $request, Devis $Devis, EntityManagerInterface $em)
     {
+        $status = $this->statusService->getStatus();
+
+        if ($status == true) {
+            return $this->render('unauthorized.html.twig');
+        }
+
         $form = $this->createForm(DevisType::class, $Devis);
         $form->handleRequest($request);
         if($form->isSubmitted() && $form->isValid()){
@@ -68,6 +87,12 @@ class DevisController extends AbstractController
     #[Route("/supprimerdevis/{id<\d+>}", name: "app_supprimer_devis")]
     public function supprimerDevis(Devis $Devis, EntityManagerInterface $em)
     {
+        $status = $this->statusService->getStatus();
+
+        if ($status == true) {
+            return $this->render('unauthorized.html.twig');
+        }
+
         $em ->remove($Devis);
         $em ->flush();
         return $this->redirectToRoute('app_devis');
